requirements.txt
kiteconnect>=4.2.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
schedule>=1.2.0
requests>=2.31.0
websocket-client>=1.6.0
python-dotenv>=1.0.0
pyotp>=2.9.0
matplotlib>=3.7.0
pytz>=2023.3
